Add a data producer
