+++ setup.cfg
@@ -31,6 +31,7 @@ packages = find:
 install_requires = 
 	click >= 8.1.0
 	faker >= 15.3.0
+	confluent-kafka >= 1.9.2
 tests_require =
 	pytest >= 7.2.0
 	pytest-mypy >= 0.10.3
